Share DataGridFilterManager across Lite/Dashboard (→ Ui)
The generic DataGridFilterManager<T> + IDataGridFilterManager were identical in
both apps except that Lite preserved the user's sort order across refresh/filter
cycles (SetItemsSourcePreservingSort) while Dashboard reassigned ItemsSource
directly. Moved one copy to PerformanceMonitor.Ui, adopting Lite's
sort-preserving superset.

BEHAVIOR CHANGE (intended, approved): Dashboard DataGrids now also preserve sort
order when data refreshes or filters change — previously a refresh dropped the
user's sort. Strict UX improvement; achieves Lite/Dashboard parity.

The shared manager calls Common.ColumnFilterMatcher.MatchesFilter directly
(instead of each app's DataGridFilterService), so it no longer depends on per-app
code. Added a Ui → Common project reference for ColumnFilterState/ColumnFilterMatcher
(no cycle: Common references no project). Both app copies deleted; the one caller
missing the Ui import (ServerTab.Filters.cs) got it.

Solution builds clean; Dashboard.Tests (488), Lite.Tests (544), Installer.Tests
(80) all pass.
